edit and delete a message has been implemented successfully

+++ zipiApi/Controller/chatRequest.js
@@ -1,7 +1,7 @@
-const { addChat, retriveChats, updateReadStatusOfOneChat, deleteOneChat, udpateOneChat }= require('../Module/chat');
+const { addChat, retriveChats, updateReadStatusOfOneChat, deleteOneChat, udpateOneChat, doesMessageExist }= require('../Module/chat');
 const { decryptToken }= require('../Middleware/JWT');
 const { successMessage, userSuccess, failureMessage, success, failure } = require('../Middleware/handleResponse');
-const { findUserByUserName, findOneUserById } = require('../Module/user');
+const { findUserByUserName, findOneUserById, doesUserExist } = require('../Module/user');
 
 const saveMessage= async(req,res,next)=>{
     try{
@@ -41,8 +41,10 @@ const updateMessage= async (req, res, next) =>{
     try{
         if(!req.params.messageId)return res.json(failureMessage("Message Id is required"));
         let activeUser= await findOneUserById(req.body.id);
+        let specificMessage= await doesMessageExist(req.params.messageId,activeUser.username);
+        if(!specificMessage)return res.json(failureMessage("Message with this Id does not exist"));
         let messageUpdated= await udpateOneChat(req.params.messageId, activeUser.username,req.body);
-        return messageUpdated?res.status(200).json(success): res.status(400).json(failure);
+        return messageUpdated?res.status(200).json(successMessage("Chat updated succesfully")): res.status(400).json(failureMessage("Your chat has not been updated yet. We are working with team to resolve this issue as soon as possible"));
     }catch(e){
         next(e);
     }
@@ -52,9 +54,10 @@ const deleteMessage= async (req, res, next)=>{
     try{
         if(!req.params.messageId)return res.json(failureMessage("Message Id is required"));
         let activeUser= await findOneUserById(req.body.id);
+        let specificMessage= await doesMessageExist(req.params.messageId,activeUser.username);
+        if(!specificMessage)return res.json(failureMessage("Message with this Id does not exist"));
         let messageDeleted= await deleteOneChat(req.params.messageId,activeUser.username);
-        return messageDeleted? res.status(200).json(success): res.status(400).json(failure);
-
+        return messageDeleted? res.status(200).json(successMessage("Chat Deleted Succesfully")): res.status(400).json(failureMessage("Chat Not Deleted"));
     }catch(e){
         next(e)
     }
